requirements.txt
Flask==2.3.3
Jinja2==3.1.3
mysqlclient==2.2.4
SQLAlchemy==2.0.30
requests==2.32.2
python-dotenv==1.0.1
flask-cors==4.0.0